moved benchmark tests around and added option to output benchmarks to ES
