+++ tests/contrib/django/django_tests.py
@@ -969,142 +969,6 @@ def test_stacktrace_filtered_for_elasticapm(client, django_elasticapm_client):
     assert spans[1]['stacktrace'][0]['module'].startswith('django.template'), spans[1]['stacktrace'][0]['function']
 
 
-@pytest.mark.parametrize('django_elasticapm_client', [{'_wait_to_first_send': 100}], indirect=True)
-def test_perf_template_render(benchmark, client, django_elasticapm_client):
-    responses = []
-    with mock.patch("elasticapm.traces.TransactionsStore.should_collect") as should_collect:
-        should_collect.return_value = False
-        with override_settings(**middleware_setting(django.VERSION,
-                                                    ['elasticapm.contrib.django.middleware.TracingMiddleware'])):
-            benchmark(lambda: responses.append(
-                client_get(client, reverse("render-heavy-template"))
-            ))
-    for resp in responses:
-        assert resp.status_code == 200
-
-    transactions = django_elasticapm_client.instrumentation_store.get_all()
-
-    # If the test falls right at the change from one minute to another
-    # this will have two items.
-    assert len(transactions) == len(responses)
-    for transaction in transactions:
-        assert len(transaction['spans']) == 2
-        assert transaction['result'] == 'HTTP 2xx'
-
-
-@pytest.mark.parametrize('django_elasticapm_client', [{'_wait_to_first_send': 100}], indirect=True)
-def test_perf_template_render_no_middleware(benchmark, client, django_elasticapm_client):
-    responses = []
-    with mock.patch(
-            "elasticapm.traces.TransactionsStore.should_collect") as should_collect:
-        should_collect.return_value = False
-        benchmark(lambda: responses.append(
-            client_get(client, reverse("render-heavy-template"))
-        ))
-    for resp in responses:
-        assert resp.status_code == 200
-
-    transactions = django_elasticapm_client.instrumentation_store.get_all()
-    assert len(transactions) == 0
-
-
-@pytest.mark.parametrize('django_elasticapm_client', [{'_wait_to_first_send': 100}], indirect=True)
-@pytest.mark.django_db(transaction=True)
-def test_perf_database_render(benchmark, client, django_elasticapm_client):
-    responses = []
-    django_elasticapm_client.instrumentation_store.get_all()
-
-    with mock.patch("elasticapm.traces.TransactionsStore.should_collect") as should_collect:
-        should_collect.return_value = False
-
-        with override_settings(**middleware_setting(django.VERSION,
-                                                    ['elasticapm.contrib.django.middleware.TracingMiddleware'])):
-            benchmark(lambda: responses.append(
-                client_get(client, reverse("render-user-template"))
-            ))
-        for resp in responses:
-            assert resp.status_code == 200
-
-        transactions = django_elasticapm_client.instrumentation_store.get_all()
-
-        assert len(transactions) == len(responses)
-        for transaction in transactions:
-            assert len(transaction['spans']) in (102, 103)
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize('django_elasticapm_client', [{'_wait_to_first_send': 100}], indirect=True)
-def test_perf_database_render_no_instrumentation(benchmark, django_elasticapm_client):
-    django_elasticapm_client.instrumentation_store.get_all()
-    responses = []
-    with mock.patch("elasticapm.traces.TransactionsStore.should_collect") as should_collect:
-        should_collect.return_value = False
-
-        client = _TestClient()
-        benchmark(lambda: responses.append(
-            client_get(client, reverse("render-user-template"))
-        ))
-
-        for resp in responses:
-            assert resp.status_code == 200
-
-        transactions = django_elasticapm_client.instrumentation_store.get_all()
-        assert len(transactions) == 0
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize('django_elasticapm_client', [{
-    '_wait_to_first_send': 100,
-    'flush_interval': 100
-}], indirect=True)
-def test_perf_transaction_with_collection(benchmark, django_elasticapm_client):
-    django_elasticapm_client.instrumentation_store.get_all()
-    with mock.patch("elasticapm.traces.TransactionsStore.should_collect") as should_collect:
-        should_collect.return_value = False
-        django_elasticapm_client.events = []
-
-        client = _TestClient()
-
-        with override_settings(**middleware_setting(django.VERSION,
-                                                    ['elasticapm.contrib.django.middleware.TracingMiddleware'])):
-            for i in range(10):
-                resp = client_get(client, reverse("render-user-template"))
-                assert resp.status_code == 200
-
-        assert len(django_elasticapm_client.events) == 0
-
-        # Force collection on next request
-        should_collect.return_value = True
-
-        @benchmark
-        def result():
-            # Code to be measured
-            return client_get(client, reverse("render-user-template"))
-
-        assert result.status_code is 200
-        assert len(django_elasticapm_client.events) > 0
-
-
-@pytest.mark.django_db
-@pytest.mark.parametrize('django_elasticapm_client', [{'_wait_to_first_send': 100}], indirect=True)
-def test_perf_transaction_without_middleware(benchmark, django_elasticapm_client):
-    django_elasticapm_client.instrumentation_store.get_all()
-    with mock.patch("elasticapm.traces.TransactionsStore.should_collect") as should_collect:
-        should_collect.return_value = False
-        client = _TestClient()
-        django_elasticapm_client.events = []
-        for i in range(10):
-            resp = client_get(client, reverse("render-user-template"))
-            assert resp.status_code == 200
-
-        assert len(django_elasticapm_client.events) == 0
-
-        @benchmark
-        def result():
-            # Code to be measured
-            return client_get(client, reverse("render-user-template"))
-
-        assert len(django_elasticapm_client.events) == 0
 
 
 @pytest.mark.skipif(django.VERSION > (1, 7),
